requirements.txt
beautifulsoup4==4.12.2
dhanhq==1.2.2
google-api-python-client==2.110.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
requests==2.31.0
